+++ src/spellbot/cogs/score_cog.py
@@ -8,6 +8,7 @@ from discord_slash.model import ContextMenuType
 
 from .. import SpellBot
 from ..interactions import ScoreInteraction
+from ..metrics import add_span_context
 from ..utils import for_all_callbacks
 
 logger = logging.getLogger(__name__)
@@ -19,15 +20,17 @@ class ScoreCog(commands.Cog):
         self.bot = bot
 
     @cog_ext.cog_context_menu(target=ContextMenuType.USER, name="View Score")
-    @tracer.wrap(name="command", resource="view_score")
+    @tracer.wrap(name="interaction", resource="view_score")
     async def view_score(self, ctx: MenuContext):
+        add_span_context(ctx)
         assert ctx.target_author
         async with ScoreInteraction.create(self.bot, ctx) as interaction:
             await interaction.execute(target=ctx.target_author)
 
     @cog_ext.cog_slash(name="score", description="View your game record on this server.")
-    @tracer.wrap(name="command", resource="score")
+    @tracer.wrap(name="interaction", resource="score")
     async def score(self, ctx: SlashContext):
+        add_span_context(ctx)
         async with ScoreInteraction.create(self.bot, ctx) as interaction:
             await interaction.execute(target=ctx.author)
 
@@ -35,8 +38,9 @@ class ScoreCog(commands.Cog):
         name="history",
         description="View historical game records in this channel.",
     )
-    @tracer.wrap(name="command", resource="history")
+    @tracer.wrap(name="interaction", resource="history")
     async def history(self, ctx: SlashContext):
+        add_span_context(ctx)
         async with ScoreInteraction.create(self.bot, ctx) as interaction:
             await interaction.history()