+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v7.10.5](https://github.com/lexicalunit/spellbot/releases/tag/v7.10.5) - 2021-12-15
+
 ### Changed
 
 - Group datadog traces under an "interaction" operation and include on_message with commands.